+++ CHANGELOG.md
@@ -5,8 +5,16 @@
 ### Changed
 
 - TODO
+
+## [0.3] - 2020-03-21
+
+### Added
+
 - django admin improved
 - tracking models
+- Producer telegram id and nick
+- Carrier model
+- Request model
 
 ## [0.2] - 2020-03-20